requirements.txt
python-socketio[asgi]
uvicorn
python-socketio
eventlet